requirements.txt
requests==2.28.1
pandas==1.3.5
fastapi==0.95.1
uvicorn==0.21.1
python-dotenv==0.21.1